+++ docs/CHANGELOG.rst
@@ -73,6 +73,9 @@ New Features:
   the latest commit from the default branch.
 * ``EDF`` folder is no longer created by default for new projects. It is now
   created only if needed when saving data from an eye tracking experiment.
+* Added a new function :func:`~klibs.KLUserInferface.mouse_clicked` to easily
+  check the input event queue for clicks and releases of mouse buttons, similar
+  to :func:`~klibs.KLUserInferface.key_pressed` for the keyboard. 
 
 
 API Changes: